Module Imap.Cmd

Commands

type 'a cmd

val encode : string -> 'a cmd -> [ `Next of string * 'x | `Wait of 'x | `End ] as 'x

val process : 'a cmd -> Response.untagged -> ('a cmd, string) Stdlib.result

val finish : 'a cmd -> 'a

val login : string -> string -> unit cmd

val logout : unit cmd

val create : string -> unit cmd

create imap name creates a mailbox named name.

val delete : string -> unit cmd

delete imap name deletes mailbox name.

val rename : string -> string -> unit cmd

rename imap oldname newname renames mailbox oldname to newname.

val noop : unit cmd

noop imap does nothing. Since any command can return a status update as untagged data, the noop command can be used as a periodic poll for new messages or message status updates during a period of inactivity.

val list : ?⁠ref:string -> string -> (Response.mailbox_flag list * char option * string) list cmd

list imap ref m returns the list of mailboxes with names matching ref.

module Status : sig ... end

val status : string -> 'a Status.t -> 'a option cmd

status imap mbox items requests status items for mailbox mbox.

val copy : Common.seq list -> string -> unit cmd

copy imap nums mbox copies messages with sequence number in nums to mailbox mbox.

val uid_copy : Common.uid list -> string -> unit cmd

val expunge : unit cmd

expunge imap permanently removes all messages that have the Deleted flag set from the currently selected mailbox.

val uid_expunge : Common.uid list -> unit cmd

Requires UIDPLUS extension.

module Search : sig ... end

val search : Search.key -> (Common.seq list * Common.modseq option) cmd

uid_search imap key returns the set of UIDs of messages satisfying the criteria key.

val uid_search : Search.key -> (Common.uid list * Common.modseq option) cmd

val examine : string -> unit cmd

select imap m selects the mailbox m for access.

val select : string -> unit cmd

select imap m selects the mailbox m for access.

val append : string -> ?⁠flags:Response.flag list -> ?⁠internaldate:string -> string -> unit cmd

append imap mbox flags id data appends data as a new message to the end of the mailbox mbox. An optional flag list can be passed using the flags argument.

module Fetch : sig ... end

val fetch : ?⁠changed_since:Common.modseq -> Common.seq list -> 'a Fetch.t -> ('a -> unit) -> unit cmd

fetch imap uid ?changed_since set att retrieves data associated with messages with sequence number in set.

If the ?changed_since argument is passed, only those messages with CHANGEDSINCE mod-sequence value at least the passed value are affected (requires the CONDSTORE extension).

val uid_fetch : ?⁠changed_since:Common.modseq -> Common.uid list -> 'a Fetch.t -> ('a -> unit) -> unit cmd

type store_mode = [

| `Add

| `Remove

| `Set

]

type _ store_kind =

| Flags : Response.flag store_kind

| Labels : string store_kind

val store : ?⁠unchanged_since:Common.modseq -> store_mode -> Common.seq list -> 'a store_kind -> 'a list -> unit cmd

store imap ?unchanged_since mode nums kind modifies kind according to mode for those message with sequence number in nums.

If ?unchanged_since is present, then only those messages with UNCHANGEDSINCE mod-sequence value at least the passed value are affected.

val uid_store : ?⁠unchanged_since:Common.modseq -> store_mode -> Common.uid list -> 'a store_kind -> 'a list -> unit cmd